third_party/blink/web_tests/platform/generic/fast/shapes/shape-outside-floats/shape-outside-floats-diamond-margin-polygon-expected.txt

X
X
X

X
X
X
PASS elementRect('l1').top is 0
PASS elementRect('l2').top is 20
PASS elementRect('l3').top is 40
PASS elementRect('l4').top is 80
PASS elementRect('l5').top is 100
PASS elementRect('l6').top is 120
PASS successfullyParsed is true

TEST COMPLETE
